Change QPointF to QPoint, to eliminate floating point precision error accumulation

+++ hornet/view_layer/grid/grid.cpp
@@ -1,4 +1,5 @@
 #include "grid.h"
+#include <QDebug>
 #include <QPainter>
 #include <QWheelEvent>
 #include <cmath>
@@ -12,21 +13,21 @@ void Grid::paintEvent(QPaintEvent *)
     QPainter painter(this);
     painter.setRenderHint(QPainter::Antialiasing, false);
 
-    const float scaledGap = gridGap * scale;
-    const float startX = std::fmod(offset.x(), scaledGap);
-    const float startY = std::fmod(offset.y(), scaledGap);
+    const double scaledGap = gridGap * scale;
+    const double startX = std::fmod(offset.x(), scaledGap);
+    const double startY = std::fmod(offset.y(), scaledGap);
 
     painter.setPen(QPen(Qt::white, 1));
 
     const int verticalLines = static_cast<int>(std::ceil((width() - startX) / scaledGap)) + 1;
     for (int i = 0; i < verticalLines; ++i) {
-        const float x = startX + i * scaledGap;
+        const double x = startX + i * scaledGap;
         painter.drawLine(QPointF(x, 0), QPointF(x, height()));
     }
 
     const int horizontalLines = static_cast<int>(std::ceil((height() - startY) / scaledGap)) + 1;
     for (int i = 0; i < horizontalLines; ++i) {
-        const float y = startY + i * scaledGap;
+        const double y = startY + i * scaledGap;
         painter.drawLine(QPointF(0, y), QPointF(width(), y));
     }
 }
